requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sqlmodel
aiosqlite
celery
alembic
sqlalchemy
greenlet
redis
black
isort
pre-commit